requirements.txt
django==1.9.4
django_model_utils==2.4
feedparser==5.2.1
python-dateutil==2.5.0